+++ src/main/java/com/world/ico/dao/ManagementFeeDao.java
@@ -3,7 +3,8 @@ package com.world.ico.dao;
 import com.world.ico.entity.ManagementFeePo;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
-import org.springframework.data.repository.query.Param;
+
+import java.math.BigDecimal;
 
 /**
  * Created by lsb on 2018/10/18.
@@ -11,6 +12,6 @@ import org.springframework.data.repository.query.Param;
 public interface ManagementFeeDao extends PagingAndSortingRepository<ManagementFeePo, Integer> {
 
     @Query(value = "SELECT management_fee from MANAGEMENT_FEE ORDER BY date limit 1", nativeQuery = true)
-    Double getManagementFee();
+    BigDecimal getManagementFee();
 
 }
